feat: Added inital Field types, refactored segments into separate modules
